# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nifty100 Financial Intelligence API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/companies/{ticker}": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Get Company",
        "operationId": "get_company_api_v1_companies__ticker__get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyBase"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/pl": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Get Pl",
        "operationId": "get_pl_api_v1_companies__ticker__pl_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfitLossResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/bs": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Get Bs",
        "operationId": "get_bs_api_v1_companies__ticker__bs_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceSheetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/cashflow": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Get Cashflow",
        "operationId": "get_cashflow_api_v1_companies__ticker__cashflow_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/ratios": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Get Ratios",
        "operationId": "get_ratios_api_v1_companies__ticker__ratios_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatiosResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/tearsheet": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Get Tearsheet",
        "operationId": "get_tearsheet_api_v1_companies__ticker__tearsheet_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TearsheetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/screener": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Screener",
        "operationId": "screener_api_v1_screener_get",
        "parameters": [
          {
            "name": "min_market_cap",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Market Cap"
            }
          },
          {
            "name": "max_market_cap",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Market Cap"
            }
          },
          {
            "name": "sector",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sector"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompanyBase"
                  },
                  "title": "Response Screener Api V1 Screener Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sectors/{sector}/companies": {
      "get": {
        "tags": [
          "Sectors"
        ],
        "summary": "Get Sector Companies",
        "operationId": "get_sector_companies_api_v1_sectors__sector__companies_get",
        "parameters": [
          {
            "name": "sector",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sector"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SectorSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/peers/{group_name}": {
      "get": {
        "tags": [
          "Peers"
        ],
        "summary": "Get Peers By Group",
        "operationId": "get_peers_by_group_api_v1_peers__group_name__get",
        "parameters": [
          {
            "name": "group_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/peers/compare": {
      "get": {
        "tags": [
          "Peers"
        ],
        "summary": "Get Company Peers",
        "operationId": "get_company_peers_api_v1_companies__ticker__peers_compare_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BalanceSheet": {
        "properties": {
          "year": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "total_assets": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Assets"
          },
          "equity_capital": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equity Capital"
          },
          "borrowings": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Borrowings"
          }
        },
        "type": "object",
        "title": "BalanceSheet"
      },
      "BalanceSheetResponse": {
        "properties": {
          "company_id": {
            "type": "string",
            "title": "Company Id"
          },
          "balance_sheets": {
            "items": {
              "$ref": "#/components/schemas/BalanceSheet"
            },
            "type": "array",
            "title": "Balance Sheets"
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "balance_sheets"
        ],
        "title": "BalanceSheetResponse"
      },
      "Cashflow": {
        "properties": {
          "year": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "operating_cash_flow": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operating Cash Flow"
          },
          "capex": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capex"
          },
          "free_cash_flow": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Free Cash Flow"
          }
        },
        "type": "object",
        "title": "Cashflow"
      },
      "CashflowResponse": {
        "properties": {
          "company_id": {
            "type": "string",
            "title": "Company Id"
          },
          "cashflows": {
            "items": {
              "$ref": "#/components/schemas/Cashflow"
            },
            "type": "array",
            "title": "Cashflows"
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "cashflows"
        ],
        "title": "CashflowResponse"
      },
      "CompanyBase": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "nse_ticker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nse Ticker"
          },
          "bse_ticker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bse Ticker"
          },
          "broad_sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Broad Sector"
          },
          "sub_sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Sector"
          },
          "market_cap_cr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap Cr"
          },
          "current_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Price"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "CompanyBase"
      },
...
      "FinancialRatio": {
        "properties": {
          "year": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "return_on_equity_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return On Equity Pct"
          },
          "debt_to_equity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Debt To Equity"
          },
          "free_cash_flow_cr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Free Cash Flow Cr"
          }
        },
        "type": "object",
        "title": "FinancialRatio"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PeerComparison": {
        "properties": {
          "company_id": {
            "type": "string",
            "title": "Company Id"
          },
          "pe_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pe Ratio"
          },
          "roe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roe"
          },
          "market_cap_cr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap Cr"
          }
        },
        "type": "object",
        "required": [
          "company_id"
        ],
        "title": "PeerComparison"
      },
      "PeerResponse": {
        "properties": {
          "group_name": {
            "type": "string",
            "title": "Group Name"
          },
          "peers": {
            "items": {
              "$ref": "#/components/schemas/PeerComparison"
            },
            "type": "array",
            "title": "Peers"
          }
        },
        "type": "object",
        "required": [
          "group_name",
          "peers"
        ],
        "title": "PeerResponse"
      },
...
      "ProfitLoss": {
        "properties": {
          "year": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "sales": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sales"
          },
          "operating_profit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operating Profit"
          },
          "net_profit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Profit"
          }
        },
        "type": "object",
        "title": "ProfitLoss"
      },
      "ProfitLossResponse": {
        "properties": {
          "company_id": {
            "type": "string",
            "title": "Company Id"
          },
          "profit_loss": {
            "items": {
              "$ref": "#/components/schemas/ProfitLoss"
            },
            "type": "array",
            "title": "Profit Loss"
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "profit_loss"
        ],
        "title": "ProfitLossResponse"
      },
      "RatiosResponse": {
        "properties": {
          "company_id": {
            "type": "string",
            "title": "Company Id"
          },
          "ratios": {
            "items": {
              "$ref": "#/components/schemas/FinancialRatio"
            },
            "type": "array",
            "title": "Ratios"
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "ratios"
        ],
        "title": "RatiosResponse"
      },
...
      "SectorSummary": {
        "properties": {
          "sector": {
            "type": "string",
            "title": "Sector"
          },
          "companies": {
            "items": {
              "$ref": "#/components/schemas/CompanyBase"
            },
            "type": "array",
            "title": "Companies"
          }
        },
        "type": "object",
        "required": [
          "sector",
          "companies"
        ],
        "title": "SectorSummary"
      },
      "TearsheetResponse": {
        "properties": {
          "company_id": {
            "type": "string",
            "title": "Company Id"
          },
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "summary"
        ],
        "title": "TearsheetResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}